+++ css-inline/src/html/attributes.rs
@@ -58,7 +58,7 @@ pub(crate) enum Cache {
 }
 
 impl Class {
-    fn new(value: StrTendril) -> Class {
+    pub(crate) fn new(value: StrTendril) -> Class {
         // Build a Bloom filter for all element's classes
         let mut cache = BloomFilter::new();
         let mut classes = value.split(SELECTOR_WHITESPACE).filter(|s| !s.is_empty());